requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.16.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.4.0
